Moved SerializedDeposit to interfaces
